Fix Controller

+++ src/main/java/com/example/Feng_Shui_Koi_Consulting_System/service/AdvertisementService.java
@@ -126,6 +126,9 @@ public class AdvertisementService {
     public AdvertisementResponse verifyAd(VerifyAdRequest request) {
         Advertisement advertisement = advertisementRepo.findById(request.getAdID())
                 .orElseThrow(() -> new AppException(ErrorCode.AD_NOT_EXIST));
+        if(!userService.getMyInfo().getRoleName().equals("ADMIN")){
+            throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
         if (!request.getNewStatus().equals("Verified") && !request.getNewStatus().equals("Rejected")) {
             throw new AppException(ErrorCode.STATUS_INVALID);
         }
@@ -159,6 +162,9 @@ public class AdvertisementService {
     public AdvertisementResponse updateAdvertisement(String adID, AdvertisementUpdateRequest request) {
         Advertisement advertisement = advertisementRepo.findById(adID)
                 .orElseThrow(() -> new AppException(ErrorCode.AD_NOT_EXIST));
+        if(!userService.getMyInfo().getUserID().equals(advertisement.getUser().getUserID())){
+            throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
         advertisementMapper.updateAdvertisement(advertisement, request, elementRepo, categoryService);
         advertisement.setStatus("Pending");
         Set<Ads_Image> currentImages = advertisement.getImagesAd();
@@ -187,6 +193,10 @@ public class AdvertisementService {
     public void deleteAdvertisement(String adID) {
         Advertisement advertisement = advertisementRepo.findById(adID)
                 .orElseThrow(() -> new AppException(ErrorCode.AD_NOT_EXIST));
+        if(!userService.getMyInfo().getUserID().equals(advertisement.getUser().getUserID())
+                || !userService.getMyInfo().getRoleName().equals("ADMIN")){
+            throw new AppException(ErrorCode.UNAUTHORIZED);
+        }
         advertisementRepo.delete(advertisement);
     }
